Fix: Situation where an old token is used for authentication
Due to our new server being in Europe and not korea, the latency is significantly higher. This bug is not new, but since because of the speed now is the first time that it is affecting us.
The situation is that, after logging in, the process `silentRefresh` is started, which will straight away refresh the auth tokens.
But since the new tokens are not so fast, the client has enough time to `navigate("/received/active");` which will send a request with the old tokens.
At that point, the server will reject the tokens because they are old and the client will default to logging out.

|client requests tokens on login --------------------------------.
                                                                 |
.----------------------------------- server sends auth tokens (A)|
|
|client starts silentRefresh
|silentRefresh requests refresh tokens----------------------------.
|(bug fix here!)                                                  |
|client requests coupon list using token (A) --------------.      |
                                                           |      |
.---------------------------------- server refreshes tokens|to (B)|
|                                                          |
|client updates tokens to (B)                              |
                                                           |
.---------server expects (B) but finds (A). Send auth error/
|
|client logs out on error